Revert to 4.1.0.3

+++ upload/system/library/cache/file.php
@@ -87,16 +87,15 @@ class File {
 	 * Destructor
 	 */
 	public function __destruct() {
-		if (mt_rand(1, 100) == 1) {
-			$files = glob(DIR_CACHE . 'cache.*');
-			if ($files) {
-				foreach ($files as $file) {
-					$time = substr(strrchr($file, '.'), 1);
+		$files = glob(DIR_CACHE . 'cache.*');
 
-					if ($time < time()) {
-						if (!@unlink($file)) {
-							clearstatcache(false, $file);
-						}
+		if ($files && mt_rand(1, 100) == 1) {
+			foreach ($files as $file) {
+				$time = substr(strrchr($file, '.'), 1);
+
+				if ($time < time()) {
+					if (!@unlink($file)) {
+						clearstatcache(false, $file);
 					}
 				}
 			}
